% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asgs.geocode-package.R
\docType{package}
\name{asgs.geocode-package}
\alias{asgs.geocode}
\alias{asgs.geocode-package}
\title{asgs.geocode: Geocode to the Australian Statistical Geography Standard}
\description{
The Australian Statistical Geography Standard (ASGS) is a hierarchical geographical classification used by the Australian Bureau of Statistics (ABS) to enable the publication of statistics that are comparable and spatially integrated. This package provides functions to geocode addresses, points and polygons to the ASGS.
}
\author{
\strong{Maintainer}: Bas Latcham \email{bas@latcham.au} (\href{https://orcid.org/0000-0002-1733-7350}{ORCID})

}
\keyword{internal}
