% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asgs_layer.r
\name{asgs_layer}
\alias{asgs_layer}
\title{Return the ASGS layer ID}
\usage{
asgs_layer(year, geo)
}
\arguments{
\item{year}{(numeric) the year of the ASGS}

\item{geo}{(character) the geographic level (eg "sa1", "SUA", "lga")}
}
\value{
The ID of the layer
}
\description{
This.
}
