% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asgs_servers.R
\name{asgs_servers}
\alias{asgs_servers}
\title{The URLs for the ASGS Servers}
\usage{
asgs_servers()
}
\value{
A named list of the URLs
}
\description{
The folders ASGS2016 and ASGS2021 contain SEARCH services. The SEARCH
services contain layers for different geographies and years (eg the 2021
folder contains 2022 LGAs).
}
