% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.r
\name{geocode}
\alias{geocode}
\title{Geocode an address using ArcGIS}
\usage{
geocode(address)
}
\arguments{
\item{address}{The address to geocode}
}
\value{
A \code{sf} object containing the address and its coordinates
}
\description{
Geocode an address using ArcGIS
}
\examples{
\dontrun{
geocode("Melbourne Town Hall")
}
}
