% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asgs.r
\name{get_asgs}
\alias{get_asgs}
\title{Get the ASGS of an \code{sf} object}
\usage{
get_asgs(.data, geo, year)
}
\arguments{
\item{.data}{(sf) The input data to geocode to ASGS}

\item{geo}{(character) the geographic level (eg "sa1", "SUA", "lga")}

\item{year}{(numeric) the year of the ASGS}
}
\value{
an \code{\link[sf:sf]{sf::sf()}} object containing the ASGS geometry and details
}
\description{
This function geocodes an \code{\link[sf:sf]{sf::sf()}} object to the Australian Statistical
Geography Standard.
}
\examples{
\dontrun{
melbourne <- geocode("Melbourne Town Hall")
get_asgs(melbourne, "sa1", 2016)
}
}
