% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers.r
\name{get_layers}
\alias{get_layers}
\title{Get ABS ASGS Layers}
\usage{
get_layers()
}
\value{
A data frame of the ABS ASGS layers
}
\description{
Get a data frame of the ABS ASGS layers
}
\examples{
get_layers()

}
