% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_servers.r
\name{get_servers}
\alias{get_servers}
\title{Get the list of servers from the ABS ArcGIS REST API}
\usage{
get_servers()
}
\value{
A character vector of the servers
}
\description{
Get the list of servers from the ABS ArcGIS REST API
}
