% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_geo.r
\name{guess_geo}
\alias{guess_geo}
\title{Guess the ID of a geographic layer}
\usage{
guess_geo(year, geo)
}
\arguments{
\item{year}{The year of the ASGS}

\item{geo}{The geographic level (eg sa1, SUA)}
}
\value{
The ID of the layer
}
\description{
Guess the ID of a geographic layer
}
