% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbc2016.R
\docType{data}
\name{mbc2016}
\alias{mbc2016}
\title{Meshblock Counts 2016}
\format{
\subsection{\code{mbc2016}}{

A data frame with 358,127 rows and 6 variables:
\describe{
\item{mb_code_2016}{Meshblock code}
\item{mb_category}{Meshblock category}
\item{area_albers_sqkm}{Area in square kilometres}
\item{dwellings}{Count of dwellings}
\item{persons}{Count of persons usually resident}
\item{state_code}{State code}
}
}
}
\source{
\href{https://www.abs.gov.au/AUSSTATS/abs@.nsf/allprimarymainfeatures/B58F9B40B7D4B4CDCA257AD0000F1EF0?opendocument=}{ABS}
}
\usage{
mbc2016
}
\description{
Meshblock counts from the 2016 Census of Population and Housing, including
category and counts of dwellings and persons by usual residence.
}
\keyword{datasets}
