% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbc2021.R
\docType{data}
\name{mbc2021}
\alias{mbc2021}
\title{Meshblock Counts 2021}
\format{
\subsection{\code{mbc2021}}{

A data frame with 368,285 rows and 15 variables:
\describe{
\item{mb_code_2021}{Meshblock code}
\item{mb_category}{Meshblock category}
\item{area_albers_sqkm}{Area in square kilometres}
\item{dwellings}{Count of dwellings}
\item{persons}{Count of persons usually resident}
\item{state_code}{State code}
\item{ASGS identifiers}{Various columns to identify the ASGS hierarchy}
}
}
}
\source{
\href{https://www.abs.gov.au/census/guide-census-data/mesh-block-counts/latest-release}{ABS}
}
\usage{
mbc2021
}
\description{
Meshblock counts from the 2021 Census of Population and Housing, including
category and counts of dwellings and persons by usual residence.
}
\keyword{datasets}
