% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_data.R
\name{osm_bind}
\alias{osm_bind}
\title{Bind the results of an OpenStreetMap (OSM) feature query}
\usage{
osm_bind(
  sf_list,
  types = c("points", "lines", "polygons", "multilines", "multipolygons"),
  crs = 4326
)
}
\arguments{
\item{sf_list}{(\code{osmdata} object) as returned by
\code{osm_query_features}}

\item{types}{(character vector) which SF types do you want. Valid options:
lines, points, polygons, multilines, multipolygons}

\item{crs}{(numeric, default = 4326) which CRS should the result be?
\code{osmdata} returns 4326 by default, but in order to intersect two sfs
they must have the same crs, and if you are dealing with Australian
geographies you should be using 7844.}
}
\value{
a \code{sf}
}
\description{
An OSM query returns a list of SFs of different types. This function binds
the ones you want together. If a requested geometry is NULL it will not be
returned, obviously. You need to use a CRS with latitude and longitude, not
decimal degrees.
}
