% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_data.R
\name{osm_find}
\alias{osm_find}
\title{Query OpenStreetMap (OSM) for features}
\usage{
osm_find(bb, feature_key, feature_values = NULL)
}
\arguments{
\item{bb}{(bounding box) the area to be queried}

\item{feature_key}{(character) the feature key to query. Feature keys can be
found here \code{osmdata::available_features()}}

\item{feature_values}{(character; default = \code{NULL}) a vector of feature values to query, passed
to \code{osmdata::add_osm_feature()}. Use \code{NULL} to return all values
of your feature.}
}
\value{
a list of \code{sf} objects, technically an \code{osmdata} object
}
\description{
Get the SF details for each requested feature. The list of OSM features is
\href{https://wiki.openstreetmap.org/wiki/Map_features}{here}. Use in
conjunction with \code{osm_bind}.
}
\examples{
\dontrun{
# Let's get all the parkland around Parramatta Park
# Geocode it to a point
address <- "Parramatta Park, Parramatta NSW"
pt <- geocode(address)

# Expand the bounding box
bbox <- pt |>
	# Transform to use meters
	sf::st_transform(crs = 7845) |>
	# 500m buffer
	sf::st_buffer(dist = 500) |>
	# OSM needs this crs
	sf::st_transform("+proj=longlat +datum=WGS84") |>
	sf::st_bbox()

# Find the parks
feats_raw <- osm_find(bbox, "leisure", "park")
# OSM returns different geometries. We want polygon-flavoured ones
feats <- osm_bind(feats_raw, c("polygons", "multipolygons"))

# View it!
mapview::mapview(feats, zcol = "name")
}
}
