% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_data.R
\name{osm_summarise}
\alias{osm_summarise}
\title{Summarise OpenStreetMap (OSM) features by group}
\usage{
osm_summarise(osm_sf, ..., .f, units = NULL)
}
\arguments{
\item{osm_sf}{(\code{sf}) SF data you want to summarise.}

\item{...}{(unquoted character; optional) columns to group by}

\item{.f}{(character) summary function. At the moment only "length", "area"
or "count". Assumes the underlying data are additive.}

\item{units}{(character, default = NULL) units to summarise to, usually "km"
or "km^2. If left blank it will use some variant of meters.}
}
\value{
a simple tibble with the grouped columns and an additional summary
column
}
\description{
Generally you will do this after querying the OSM features, binding the
results, intersecting with an irregular polygon (e.g. a burnscar).
}
\details{
Requires the geometry column to be called geometry (and not geos or
some-such).
}
