% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_to_asgs.r
\name{point_to_asgs}
\alias{point_to_asgs}
\title{Geocode a point to ASGS}
\usage{
point_to_asgs(.data, layer)
}
\arguments{
\item{.data}{The input data}

\item{layer}{The layer to use}
}
\value{
ASGS data
}
\description{
Geocode a point to ASGS
}
\examples{
\dontrun{
melbourne <- geocode("Melbourne Town Hall")
point_to_asgs(melbourne)
}
}
