% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.r
\name{geocode}
\alias{geocode}
\title{Geocode an address using ArcGIS}
\usage{
geocode(address)
}
\arguments{
\item{address}{The address to geocode}
}
\value{
A one-row \code{\link[sf:sf]{sf::sf()}} object containing the address and its coordinates
}
\description{
This function geocodes an address using the ArcGIS API. The address is
suffixed with the country code "AUS", and only the top match is returned. The
more specific the address, the more likely the correct result will be
returned.
}
\examples{
\dontrun{
geocode(address = "Melbourne Town Hall")
}
}
