% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.r
\name{geocode}
\alias{geocode}
\title{Geocode an address using ArcGIS}
\usage{
geocode(address, max_results = 5L)
}
\arguments{
\item{address}{(character) The address to geocode}

\item{max_results}{(numeric; default = \code{5L}) How many address candidates to return.}
}
\value{
A one-row \code{\link[sf:sf]{sf::sf()}} object containing the address and its coordinates.
}
\description{
This function geocodes an address using the ArcGIS API. The address is
suffixed with the country code "AUS". The more specific the address, the more
likely the correct result will be returned. See details for more information.
}
\details{
The returned object has the matched address as well as columns indicating the
accuracy of the match. These columns are:
\describe{
\item{rank}{A numerical text value or floating-point value in a number field,
depending on the locator that indicates the importance of a result relative
to other results with the same name. For example, there are cities in France
and Texas named Paris. Paris, France, has a greater population than Paris,
Texas, so it has a higher rank. The smaller numbers represent higher-ranked
features. Rank is used to sort results for ambiguous queries such as Lincoln,
in which no additional information (state) is available. Rank values are
based on population or feature type.}
\item{score}{The match score of the candidate to which the address was
matched. The score can be in a range of 0 to 100, in which 100 indicates the
candidate is a perfect match.}
\item{match_status}{The status of the address match. The status can be one of
the following: matched, tied, unmatched.}
\item{type}{The feature type for results returned by a search. The Type field
only includes a value for candidates with Addr_type = POI or Locality, as an
example, for Starbucks, \verb{Type = Coffee Shop}.}
\item{address_type}{The type of address that was matched. There are many
address types, with \code{Subaddress} generally being the most spatially accurate. See
\href{https://pro.arcgis.com/en/pro-app/latest/help/data/geocoding/what-is-included-in-the-geocoded-results-.htm}{here}
(search \code{Addr_type}) for more details.}
}
}
\examples{
\dontrun{
geocode(address = "Melbourne Town Hall")
geocode(address = "Melbourne Town Hall", 2)
}
}
