% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.r
\name{geocode}
\alias{geocode}
\title{Geocode an address using ArcGIS}
\usage{
geocode(address, candidate = 1L)
}
\arguments{
\item{address}{(character) The address to geocode}

\item{candidate}{(integer) The address candidate to return. Defaults to 1
(the most likely result). Changing this can be useful if geocoding returns a
tie.}
}
\value{
A one-row \code{\link[sf:sf]{sf::sf()}} object containing the address and its coordinates.
}
\description{
This function geocodes an address using the ArcGIS API. The address is
suffixed with the country code "AUS". The more specific the address, the more
likely the correct result will be returned. See details for more information.
}
\details{
The returned object has the matched address as well as columns indicating the
accuracy of the match. These columns are:
\describe{
\item{candidate}{Which address candidate was returned. The default is 1, which is the most likely result.}
\item{score}{The match score of the candidate to which the address was
matched. The score can be in a range of 0 to 100, in which 100 indicates the
candidate is a perfect match.}
\item{match_status}{The status of the address match. The status can be one of
the following: matched, tied, unmatched.}
\item{address_type}{The type of address that was matched. There are many
address types, with \code{Subaddress} generally being the most spatially accurate. See
\href{https://pro.arcgis.com/en/pro-app/latest/help/data/geocoding/what-is-included-in-the-geocoded-results-.htm}{here}
(search \code{Addr_type}) for more details.}
}
}
\examples{
\dontrun{
geocode(address = "Melbourne Town Hall")
geocode(address = "Melbourne Town Hall", 2)
}
}
