% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asgs.r
\name{get_asgs}
\alias{get_asgs}
\title{Geocode a point to ASGS}
\usage{
get_asgs(.data, year, geo)
}
\arguments{
\item{.data}{(sf) The input data to geocode to ASGS}

\item{year}{(numeric) the year of the ASGS}

\item{geo}{(character) the geographic level (eg "sa1", "SUA", "lga")}

\item{layer}{(numeric; default = \code{NULL}) The layer to use if you know the
code. View all available layers with \code{\link[=get_layers]{get_layers()}}.}
}
\value{
an \code{\link[sf:sf]{sf::sf()}} object containing the ASGS geometry and details
}
\description{
This function geocodes an \code{\link[sf:sf]{sf::sf()}} object to the Australian Statistical Geography Standardf.
}
\examples{
\dontrun{
melbourne <- geocode("Melbourne Town Hall")
get_asgs(melbourne, 2016, "sa1")
}
}
